MCP_Client/requirements.txt
mcp>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
